Add files via upload
Armstrong is a Program that prints all the desired armstrong numbers upto 10000.

An integer with three digits is considered to be an Armstrong number if the sum of its digits' cubes is an integer. 
Taking 3**3 + 7**3 + 1**3 as an example, 371 is an Armstrong number. 

similarly if we consider four digit number 
9474 = 9**4 + 4**4 + 7**4 + 4**4 
So it is considered as four digit armstrong number